* release for fastrpc3 * fix many warnings (not so all :( ) * fix packaging problems for lenny
git-svn-id: https://fastrpc.svn.sourceforge.net/svnroot/fastrpc/python/trunk@176 b39e363e-0249-4afc-b8ac-0517a591d126

+++ pythonbuilder.h
@@ -20,7 +20,7 @@
  * Radlicka 2, Praha 5, 15000, Czech Republic
  * http://www.seznam.cz, mailto:fastrpc@firma.seznam.cz
  *
- * $Id: pythonbuilder.h,v 1.5 2008-04-01 13:19:10 burlog Exp $
+ * $Id: pythonbuilder.h,v 1.6 2008-11-14 10:18:22 burlog Exp $
  *
  * AUTHOR      Vaclav Blazek <blazek@firma.seznam.cz>
  *
@@ -166,7 +166,7 @@ public:
             {
 
                 bool utf8 = false;
-                for(long i = 0; i < memberName.size(); i++)
+                for(size_t i = 0; i < memberName.size(); i++)
                 {
                     if(memberName[i] & 0x80)
                     {
